+++ app/client/src/components/autoHeightOverlay/AutoHeightLimitHandleGroup.tsx
@@ -1,11 +1,7 @@
-import React, { useRef } from "react";
+import React from "react";
 import styled from "styled-components";
-import AutoHeightLimitHandleBorder from "./ui/AutoHeightLimitHandleBorder";
-import { useDrag } from "react-use-gesture";
-import { OVERLAY_COLOR } from "./constants";
-import { heightToRows } from "./utils";
-import AutoHeightLimitHandleLabel from "./ui/AutoHeightLimitHandleLabel";
 import { onDragCallbacksProps, onMouseHoverCallbacksProps } from "./types";
+import AutoHeightLimitHandle from "./AutoHeightLimitHandle";
 
 const AutoHeightLimitHandleGroupContainer = styled.div`
   position: absolute;
@@ -30,118 +26,6 @@ interface AutoHeightLimitHandleGroupProps {
   onMaxLimitMouseHoverCallbacks: onMouseHoverCallbacksProps;
   onMinLimitMouseHoverCallbacks: onMouseHoverCallbacksProps;
 }
-
-interface AutoHeightLimitHandleContainerProps {
-  height: number;
-}
-
-const AutoHeightLimitHandleContainer = styled.div<
-  AutoHeightLimitHandleContainerProps
->`
-  position: absolute;
-  display: flex;
-  align-items: center;
-  width: 100%;
-  height: 6px;
-  transform: translateY(${(props) => props.height}px);
-`;
-
-interface AutoHeightLimitHandleDotProps {
-  isActive: boolean;
-  isDragging: boolean;
-}
-
-const AutoHeightLimitHandleDot = styled.div<AutoHeightLimitHandleDotProps>`
-  align-self: start;
-  position: absolute;
-  left: 50%;
-  top: -3px;
-  cursor: ns-resize;
-  border-radius: 50%;
-  width: 7px;
-  height: 7px;
-  transform: translate(-50%)
-    scale(${(props) => (props.isDragging ? "1.67" : "1")});
-  border: 1px solid ${OVERLAY_COLOR};
-  background-color: ${OVERLAY_COLOR};
-  box-shadow: 0px 0px 0px 2px white;
-`;
-
-interface AutoHeightLimitHandleProps {
-  cypressDataID: string;
-  height: number;
-  isActive: boolean;
-  isColliding: boolean;
-  isDragging: boolean;
-  label: string;
-  onDragCallbacks: onDragCallbacksProps;
-  onMouseHoverFunctions: onMouseHoverCallbacksProps;
-}
-
-const AutoHeightLimitHandle = ({
-  cypressDataID,
-  height,
-  isActive,
-  isColliding,
-  isDragging,
-  label,
-  onDragCallbacks,
-  onMouseHoverFunctions,
-}: AutoHeightLimitHandleProps) => {
-  const ref = useRef<HTMLDivElement>(null);
-  const { onStart, onStop, onUpdate } = onDragCallbacks;
-
-  const bind = useDrag(
-    (state) => {
-      if (state.first) {
-        onStart();
-        return;
-      }
-
-      if (state.last) {
-        onStop();
-        return;
-      }
-      const [mx, my] = state.movement;
-
-      onUpdate(mx, my);
-    },
-    { axis: "y" },
-  );
-
-  const bindings = bind();
-
-  return (
-    <AutoHeightLimitHandleContainer
-      data-cy={cypressDataID}
-      height={height}
-      ref={ref}
-      {...bindings}
-      onDragStart={(e) => {
-        e.preventDefault();
-        e.stopPropagation();
-      }}
-      onMouseDown={(e) => {
-        e.preventDefault();
-        e.stopPropagation();
-        bindings?.onMouseDown && bindings.onMouseDown(e);
-      }}
-      onMouseUp={(e) => {
-        e.preventDefault();
-        e.stopPropagation();
-      }}
-      {...onMouseHoverFunctions}
-    >
-      <AutoHeightLimitHandleBorder isActive={isActive} />
-      <AutoHeightLimitHandleDot isActive={isActive} isDragging={isDragging} />
-      {!isColliding ? (
-        <AutoHeightLimitHandleLabel isActive={isActive}>
-          {label}: {heightToRows(height)} rows
-        </AutoHeightLimitHandleLabel>
-      ) : null}
-    </AutoHeightLimitHandleContainer>
-  );
-};
 
 const AutoHeightLimitHandleGroup: React.FC<AutoHeightLimitHandleGroupProps> = ({
   isMaxDotActive,
